+++ contracts/interfaces/IEmissionSetter.sol
@@ -3,5 +3,5 @@
 pragma solidity 0.8.13;
 
 interface IEmissionSetter {
-    function getRate() external view returns (uint256 rate);
+    function getRate() external view returns (uint rate);
 }
